+++ common/src/fn_network/network_read.c
@@ -101,11 +101,21 @@ int16_t network_read(char *devicespec, uint8_t *buf, uint16_t len)
         r = network_status(devicespec, &fn_network_bw, &fn_network_conn, &fn_network_error); // TODO: Status return needs fixing.
 #endif
 
-        // check if the status failed
-        if (r != 0) return -r;
+        // check if the status failed. The buffer may be partially filled, up to client if they want to use any of it. The count is in fn_bytes_read
+        if (r != 0) {
+            fn_bytes_read = total_read;
+            // r is the FN_ERR code
+            return -r;
+        }
 
         // EOF hit, exit reading
         if (fn_network_error == 136) break;
+
+        // is there another error? The buffer may be partially filled, up to client if they want to use any of it. The count is in fn_bytes_read
+        if (fn_network_error != 1) {
+            fn_bytes_read = total_read;
+            return -FN_ERR_IO_ERROR;
+        }
 
         // we are waiting for bytes to become available while still connected.
         // Causes tight loop if there's a long delay reading from network into FN, so we may see lots of status requests