+++ Changelog.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [4.5.3] - 2024-08-32
+
+- network_read and network_read_nb will exit if there is a general error.
+  network_read will set fn_bytes_read to the bytes read into the buffer so far, for client to decide what to do.
+
 ## [4.5.2] - 2024-08-25
 
 - [atari] fuji_read_appkey no longer uses malloc, but requires the data buffer passed in to be at least 2 bytes larger than the keysize to work.